Re-organize.
