+++ lib/rules-en.js
@@ -1,4 +1,4 @@
-const { FunctionRule, RemoveRule } = require('./definitions');
+const { FunctionRule, RemoveRule } = require('./rules');
 const normalize = require('./normalize');
 
 // These two lists contain all ASCII symbols, where symbol is defined as a
