+++ lib/rules-es.js
@@ -1,5 +1,5 @@
-const en = require('./en');
-const { RemoveRule } = require('./definitions');
+const en = require('./rules-en');
+const { RemoveRule } = require('./rules');
 
 // See https://www.spanishdict.com/guide/spanish-punctuation
 const COMMON_PUNCTUATION_STR = en.COMMON_PUNCTUATION_STR + `¿¡«»`;
