+++ lib/rules.js
@@ -1,11 +1,98 @@
+const isString = require('is-string');
+
 const {
+  mapFromString,
+  isMap,
+  isFunction,
   replaceAll,
   assert,
-  Rule,
-  FunctionRule,
-  MapRule,
-} = require('./definitions');
+} = require('./helpers');
 const normalize = require('./normalize');
+
+class Rule {
+  constructor(opts) {
+    this.opts = opts;
+    // Because optimizations and browsers may not preserve
+    // instance.constructor.name, we cannot rely on it for discovering type.
+    this.type = 'Rule';
+    this.name = opts.name || 'unnamed';
+  }
+}
+
+Rule.from = (candidateRule, givenName) => {
+  if (candidateRule instanceof Rule || /Rule/.test(candidateRule.type))
+    return candidateRule;
+  let name;
+  if (givenName) {
+    name = givenName;
+  } else if (candidateRule.name) {
+    name = candidateRule.name;
+  } else if (typeof candidateRule === 'string') {
+    name = candidateRule;
+  }
+
+  let newRule;
+  if (isString(candidateRule)) {
+    newRule = new MapRule(mapFromString(candidateRule), { name });
+  } else if (isMap(candidateRule)) {
+    newRule = new MapRule(candidateRule, { name });
+  } else if (isFunction(candidateRule)) {
+    newRule = new FunctionRule(candidateRule, { name });
+  } else {
+    throw new Error(`expected a String, Function, or Map: ${candidateRule}`);
+  }
+  return newRule;
+};
+
+class FunctionRule extends Rule {
+  constructor(fn, opts = { modifiesCharacters: true }) {
+    if (!opts.name) opts.name = fn.toString();
+    super(opts);
+    this.fn = fn;
+    this.type = 'FunctionRule';
+  }
+
+  apply(s1, s2) {
+    return this.fn(s1, s2);
+  }
+}
+
+// TODO: ideally, MapRule would mixin Map.
+class MapRule extends Rule {
+  constructor(map, opts = { modifiesCharacters: false }) {
+    if (!opts.name) opts.name = map.toString();
+    super(opts);
+    assert(
+      map instanceof Map,
+      `${map} is not an instance of Map (is ${map.constructor})`
+    );
+    this.map = map;
+    this.type = 'MapRule';
+  }
+
+  apply() {
+    throw new Error('MapRules cannot be directly applied.');
+  }
+
+  entries() {
+    return this.map.entries();
+  }
+
+  keys() {
+    return this.map.keys();
+  }
+}
+
+class RemoveRule extends MapRule {
+  constructor(str, opts = { modifiesCharacters: false }) {
+    assert(
+      typeof str === 'string',
+      `${str} is not a string (is ${str.constructor})`
+    );
+    const map = mapFromString(str);
+    super(map, opts);
+  }
+}
 
 const identity = (s1, s2) => [s1, s2];
 const identityRule = new FunctionRule(identity, { name: 'identity' });
@@ -90,11 +177,6 @@ const acuteAccentRule = codepointRuleGenerator(
 
 const umlautRule = codepointRuleGenerator([umlautCombining], 'umlaut');
 
-const combiningDiacriticsBlockRule = codepointRuleGenerator(
-  combiningDiacriticsBlock,
-  'combining diacritics block'
-);
-
 const combiningDiacriticsBlockExceptAcuteAndUmlautRule = codepointRuleGenerator(
   combiningDiacriticsBlock
     .filter(codepoint => codepoint !== acuteAccentCombining)
@@ -111,32 +193,19 @@ const wordPrefixRuleCreator = word =>
     { name: `wordPrefixRule '${word}'` }
   );
 
-const en = require('./en');
-const enExport = Object.keys(en).reduce((accum, k) => {
-  accum[k] = Rule.from(en[k]);
-  return accum;
-}, {});
-
-const es = require('./es');
-const esExport = Object.keys(es).reduce((accum, k) => {
-  accum[k] = Rule.from(es[k]);
-  return accum;
-}, {});
-
 module.exports = {
   assert,
   Rule,
   MapRule,
+  FunctionRule,
+  RemoveRule,
   identityRule,
-  en: enExport,
-  es: esExport,
   CAPITALIZATION: capitalizationRule,
   UNICODE_NORMALIZATION: unicodeNormalizationRule,
   WHITESPACE_DIFFERENCES: whitespaceDifferencesRule,
   COMMON_DIACRITICS: commonDiacriticsRule,
   ACUTE_ACCENT: acuteAccentRule,
   UMLAUT: umlautRule,
-  COMBINING_DIACRITICS_BLOCK: combiningDiacriticsBlockRule,
   COMBINING_DIACRITICS_BLOCK_EXCEPT_ACUTE_AND_UMLAUT: combiningDiacriticsBlockExceptAcuteAndUmlautRule,
   wordPrefix: wordPrefixRuleCreator,
 };
